+++ temba/contacts/tests.py
@@ -2869,7 +2869,7 @@ class ContactTest(TembaTest):
                 ContactGroup.create_dynamic(self.org, self.admin, "Age field is invalid", 'age < "age"')
 
         # when creating a new contact we should only reevaluate 'empty age field' and 'urn group' groups
-        with self.assertNumQueries(33):
+        with self.assertNumQueries(30):
             contact = Contact.get_or_create_by_urns(self.org, self.admin, name="Željko", urns=["twitter:helio"])
 
         self.assertCountEqual(
@@ -3693,7 +3693,7 @@ class ContactTest(TembaTest):
         EventFire.update_campaign_events(self.campaign)
 
         # should have seven fires, one for each campaign event
-        self.assertEqual(7, EventFire.objects.all().count())
+        self.assertEqual(7, EventFire.objects.filter(event__is_active=True).count())
 
         # visit a contact detail page as a user but not belonging to this organization
         self.login(self.user1)